creo que esta todo ok

+++ xoshiro256plus.h
@@ -1,5 +1,4 @@
 #include <stdint.h>
 
-double optimized_random_probability(void);
-
+float optimized_random_probability(void);
 void seed(uint64_t seed);
